requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
requests
celery
redis
google-generativeai
pydantic-settings
pydantic[email]